Use bigger buckets in tabular data splitting
This is somewhat arbitrary, but the idea is to balance the
tradeoff between overhead starting the processing pipeline
and granularity for parallel processing.

You want coarse-grained to avoid the overheads.
But you want fine-grained to maximise exploitation (ie.
you want all cores working, instead of having some be
idle because they finished their smaller-sized chunks
early)